requirements.txt
absl-py==0.8.1
astor==0.8.0
audioread==2.1.8
bokeh==2.2.3
certifi==2020.6.20
cffi==1.13.0
chardet==3.0.4
cycler==0.10.0
decorator==4.4.0
gast==0.2.2
google-pasta==0.1.7
grpcio==1.24.1
h5py==2.10.0
holoviews==1.13.5
idna==2.10
imageio==2.6.1
imagen==2.1.0
Jinja2==2.11.2
joblib==0.14.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
librosa==0.7.1
llvmlite==0.30.0
lmdb==0.98
Markdown==3.1.1
MarkupSafe==1.1.1
matplotlib==3.1.1
networkx==2.4
numba==0.46.0
numpy==1.17.2
opencv-python==4.3.0.38
opt-einsum==3.1.0
packaging==20.4
pandas==0.25.3
panel==0.10.1
param==1.10.0
Pillow==8.0.1
protobuf==3.10.0
pyarrow==1.0.0
pycparser==2.19
pyct==0.4.8
pyparsing==2.4.2
python-dateutil==2.8.0
pytz==2019.3
pyviz-comms==0.7.6
PyWavelets==1.0.3
PyYAML==5.3.1
requests==2.24.0
resampy==0.2.2
scikit-image==0.16.1
scikit-learn==0.21.3
scipy==1.3.1
six==1.12.0
SoundFile==0.10.2
tensorboard==2.0.0
tensorflow==2.0.0
tensorflow-estimator==2.0.0
termcolor==1.1.0
--find-links https://download.pytorch.org/whl/torch_stable.html
torch==1.3.0
torchtoolbox==0.1.5
torchvision==0.4.1
tornado==6.1
tqdm==4.48.2
typing-extensions==3.7.4.3
urllib3==1.25.11
Werkzeug==0.16.0
wrapt==1.11.2